useMarkdownInternalHandlerを全ての場所で適用
